feat: finished education section with working checkbox useState
my main issue with working with these new inputs is trying to
keep the inputs in it's state when toggling between
submission and editing mode. The checkbox wasn't staying checked
if the user submitted the checkbox in the checked state, then
going back and editing the checkbox which would re-render.
I found out that setting the checked attribute on the
input tag with a boolean value toggled the checkbox
state manually, so after making that change, it worked.