feat: added a final thank you message after submission and fixed inline styling issue
I imported the finalCV.jsx into the CVform.jsx file
to re-render the page after submission with a useState,
but I was having trouble setting an inline
style with the <main> tag, but then I realized
that you have to define the style attribute with
an object with css within the javascript curly
bracket jsx. Like this: style={{css-style: "thing"}}.
This let me overwrite the background CSS style and
fixed the re-rendering issue.